+++ src/pages/auth/verify-email/index.ts
@@ -1,0 +1,1 @@
+export { default as EmailConfirmPage } from './ui/EmailConfirmPage.async'
