+++ src/pages/auth/verify-email/ui/EmailConfirmPage.async.tsx
@@ -1,0 +1,10 @@
+import { Suspense } from 'react'
+import { EmailConfirmPage } from './emailConfirmPage'
+
+const EmailConfirmPageAsync = () => (
+  <Suspense>
+    <EmailConfirmPage />
+  </Suspense>
+)
+
+export default EmailConfirmPageAsync
