+++ src/pages/auth/verify-email/ui/verify-email.module.scss
@@ -1,0 +1,13 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  align-items: center;
+  justify-content: center;
+}
+
+.button {
+  min-width: 180px;
+  margin-top: 54px;
+  margin-bottom: 72px;
+}
