--- src/pages/auth/verify-email/verify-email.module.scss
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  gap: 20px;
-  margin-top: 35px;
-}
-
-.btnContainer {
-  margin-top: 54px;
-  margin-bottom: 72px;
-}
-
-.button {
-  min-width:180px;
-}
